refactor(server): decompose _response_budget.py into 5-file package
Splits the 1301-line server._response_budget module into a directory
package with 4 sibling submodules + __init__.py facade:

- _primitives.py (~122 lines) - RESPONSE_* constants, primitive helpers
- _projection.py (~670 lines) - projection + spill-for-delivery-bound
- _spill.py (~164 lines) - bounded failure + spill envelope machinery
- _enforce.py - enforce_response_budget + checkpoint-segmented shaping
- __init__.py (~32 lines) - re-exports all 11 symbols from original __all__

The __init__.py facade preserves the flat import surface so consumer
imports keep working without modification. Cross-submodule dependencies
are routed through explicit imports; no circular-import risk.

This is part of issue #4663's decomposition to bring all source files
under the 750-line per-file ceiling enforced by
test_no_src_module_exceeds_line_limit().
