+++ src/autoskillit/server/_response_budget/_projection.py
@@ -1,134 +1,38 @@
-"""Lossless, shape-preserving enforcement for MCP handler responses."""
-
 from __future__ import annotations
 
 import hashlib
 import json
-import uuid
-from contextlib import suppress
 from pathlib import Path
 from typing import TYPE_CHECKING, Any
 
 from autoskillit.core import (
-    ASCII_YAML_POLICY,
     CLIENT_CHARS_PER_TOKEN_POLICY,
-    RESPONSE_BACKSTOP_EXEMPTION_REGISTRY,
     TokenLimit,
-    Utf8ByteLimit,
     atomic_write,
     get_logger,
 )
-from autoskillit.server._recipe_segment_delivery import (
-    RECIPE_SEGMENT_MAX_BYTES,
-    RecipeSegmentDeliveryError,
-    build_post_effect_segment_failure,
+from autoskillit.server._response_budget._primitives import (
+    RESPONSE_SPILL_METADATA_KEY,
+    _bounded_tool_name,
+    _emit_response_budget_event,
+    _minimal_same_type,
+    _preview_string,
+    _ProjectionNonconvergentError,
+)
+from autoskillit.server._response_budget._spill import (
+    _artifact_path,
+    _finalize_envelope,
+    _plain_spill_envelope,
+    _serialized_string_prefix_length,
+    _spill_metadata,
+    _total_omissions,
+    bounded_response_budget_failure,
 )
 
 if TYPE_CHECKING:
     from autoskillit.config import OutputBudgetConfig
 
 logger = get_logger(__name__)
-
-RESPONSE_SPILL_METADATA_KEY = "_autoskillit_response_spill"
-RESPONSE_SPILL_SCHEMA_VERSION = 1
-RESPONSE_SPILL_METADATA_KEYS = frozenset(
-    {
-        "schema_version",
-        "artifact_path",
-        "sha256",
-        "original_utf8_bytes",
-        "projected_utf8_bytes",
-        "omitted_chars",
-        "omitted_items",
-        "reason",
-    }
-)
-RESPONSE_SPILL_REASONS = frozenset(
-    {"oversized_values", "minimal_projection", "plain_text", "delivery_bound"}
-)
-RESPONSE_SPILL_SCHEMA_DIGEST = hashlib.sha256(
-    json.dumps(
-        {
-            "metadata_key": RESPONSE_SPILL_METADATA_KEY,
-            "metadata_keys": sorted(RESPONSE_SPILL_METADATA_KEYS),
-            "reasons": sorted(RESPONSE_SPILL_REASONS),
-            "schema_version": RESPONSE_SPILL_SCHEMA_VERSION,
-        },
-        sort_keys=True,
-        separators=(",", ":"),
-    ).encode()
-).hexdigest()
-
-RESPONSE_BUDGET_FAILURE_CAUSES = frozenset(
-    {
-        "context_unavailable",
-        "artifact_publication_failed",
-        "serialization_failed",
-        "projection_nonconvergent",
-        "irreducible_shape",
-        "schema_nonconforming",
-        "exemption_ceiling_exceeded",
-        "internal_invariant_failed",
-    }
-)
-
-
-class _ProjectionNonconvergentError(RuntimeError):
-    pass
-
-
-def _canonical_json(value: Any) -> str:
-    return json.dumps(value, ensure_ascii=False, separators=(",", ":"))
-
-
-def _estimated_tokens(original_size: int) -> int:
-    """Estimate tokens through the explicit ASCII-YAML conversion policy.
-
-    Uses the general output token limit as a coarse
-    transport-layer estimate, not a tokenizer count. Used to compare
-    payload size against ``selected_result_token_limit``.
-    """
-    if original_size == 0:
-        return 0
-    return ASCII_YAML_POLICY.to_tokens(Utf8ByteLimit(original_size)).value
-
-
-def _serialized(value: Any) -> str:
-    if isinstance(value, str):
-        return value
-    return _canonical_json(value)
-
-
-def _bounded_tool_name(tool_name: str) -> str:
-    return tool_name.encode("ascii", "replace").decode("ascii")[:64]
-
-
-def _emit_response_budget_event(event: str, **payload: Any) -> None:
-    with suppress(Exception):
-        logger.info(event, **payload)
-
-
-def emit_response_budget_failure(tool_name: str, cause: str, original_utf8_bytes: int) -> None:
-    if cause not in RESPONSE_BUDGET_FAILURE_CAUSES:
-        cause = "internal_invariant_failed"
-    _emit_response_budget_event(
-        "response_budget_failure",
-        tool_name=_bounded_tool_name(tool_name),
-        cause=cause,
-        original_utf8_bytes=original_utf8_bytes,
-    )
-
-
-def _preview_string(value: str, limit: int) -> tuple[str, int]:
-    if len(value) <= limit:
-        return value, 0
-    if limit <= 16:
-        return value[:limit], len(value) - limit
-    marker = "…[omitted]…"
-    available = max(0, limit - len(marker))
-    head = available // 2
-    tail = available - head
-    return value[:head] + marker + (value[-tail:] if tail else ""), len(value) - available
 
 
 def _project_value(value: Any, limit: int) -> tuple[Any, int, int]:
@@ -167,127 +71,6 @@ def _project_value(value: Any, limit: int) -> tuple[Any, int, int]:
         return projected_dict, omitted_chars, omitted_items
     text, omitted = _preview_string(str(value), limit)
     return text, omitted, 0
-
-
-def _minimal_same_type(value: Any) -> Any:
-    if isinstance(value, str):
-        return ""
-    if isinstance(value, list):
-        return []
-    if isinstance(value, dict):
-        return {}
-    if value is None or isinstance(value, (bool, int, float)):
-        return value
-    return ""
-
-
-def _bounded_failure(
-    *,
-    reason: str,
-    tool_name: str,
-    max_bytes: int,
-    artifact_path: str | None = None,
-) -> str:
-    payload: dict[str, Any] = {
-        "success": False,
-        "error": reason,
-        "tool_name": tool_name,
-    }
-    if artifact_path is not None:
-        payload["artifact_path"] = artifact_path
-    candidates = (
-        _canonical_json(payload),
-        '{"success":false,"error":"response_budget_failure"}',
-        '{"success":false}',
-        "{}",
-        "",
-    )
-    return next(
-        candidate for candidate in candidates if len(candidate.encode("utf-8")) <= max_bytes
-    )
-
-
-def bounded_response_budget_failure(
-    result: Any,
-    *,
-    cause: str,
-    tool_name: str,
-    max_bytes: int,
-    original_utf8_bytes: int,
-    artifact_path: str | None = None,
-) -> Any:
-    emit_response_budget_failure(tool_name, cause, original_utf8_bytes)
-    rendered = _bounded_failure(
-        reason=f"response_budget_{cause}",
-        tool_name=_bounded_tool_name(tool_name),
-        max_bytes=max_bytes,
-        artifact_path=artifact_path,
-    )
-    if isinstance(result, str):
-        return rendered
-    try:
-        return json.loads(rendered)
-    except ValueError:
-        return {"success": False, "error": "response_budget_failure"}
-
-
-def _artifact_path(artifact_dir: Path, tool_name: str) -> Path:
-    safe_name = "".join(c if c.isalnum() or c in "._-" else "_" for c in tool_name)
-    return artifact_dir / f"{safe_name or 'response'}_{uuid.uuid4().hex[:8]}.log"
-
-
-def _finalize_envelope(envelope: dict[str, Any], *, max_bytes: int) -> str:
-    metadata = envelope.get(RESPONSE_SPILL_METADATA_KEY)
-    if not isinstance(metadata, dict) or "projected_utf8_bytes" not in metadata:
-        raise _ProjectionNonconvergentError("spill metadata missing projected byte field")
-    max_decimal_width = len(str(max(0, max_bytes)))
-    seen: set[tuple[int, int]] = set()
-    for _ in range(max_decimal_width + 2):
-        rendered = _canonical_json(envelope)
-        measured = len(rendered.encode("utf-8"))
-        current = metadata.get("projected_utf8_bytes")
-        if current == measured:
-            return rendered
-        state = (current if isinstance(current, int) else -1, measured)
-        if state in seen:
-            break
-        seen.add(state)
-        metadata["projected_utf8_bytes"] = measured
-    raise _ProjectionNonconvergentError(
-        "projected byte fixed point did not converge: "
-        f"measured={len(_canonical_json(envelope).encode('utf-8'))} "
-        f"projected={metadata.get('projected_utf8_bytes')!r} "
-        f"max_bytes={max_bytes} attempted_states={len(seen)}"
-    )
-
-
-def _spill_metadata(
-    metadata: dict[str, Any],
-    *,
-    reason: str,
-    omitted_chars: int,
-    omitted_items: int,
-) -> dict[str, Any]:
-    return {
-        "schema_version": RESPONSE_SPILL_SCHEMA_VERSION,
-        **metadata,
-        "projected_utf8_bytes": 0,
-        "omitted_chars": omitted_chars,
-        "omitted_items": omitted_items,
-        "reason": reason,
-    }
-
-
-def _total_omissions(value: Any) -> tuple[int, int]:
-    if isinstance(value, str):
-        return len(value), 0
-    if isinstance(value, list):
-        nested = [_total_omissions(item) for item in value]
-        return sum(chars for chars, _ in nested), len(value) + sum(items for _, items in nested)
-    if isinstance(value, dict):
-        nested = [_total_omissions(item) for item in value.values()]
-        return sum(chars for chars, _ in nested), len(value) + sum(items for _, items in nested)
-    return 0, 0
 
 
 def _project_json_object(
@@ -383,34 +166,6 @@ def _project_json_object(
     return rendered if len(rendered.encode("utf-8")) <= max_bytes else None
 
 
-def _plain_spill_envelope(
-    original: str,
-    *,
-    metadata: dict[str, Any],
-    max_bytes: int,
-    inline_chars: int,
-    reason: str = "plain_text",
-) -> str | None:
-    preview_limit = max(0, min(inline_chars, max_bytes // 3))
-    while True:
-        preview, omitted_chars = _preview_string(original, preview_limit)
-        envelope: dict[str, Any] = {
-            RESPONSE_SPILL_METADATA_KEY: _spill_metadata(
-                metadata,
-                reason=reason,
-                omitted_chars=omitted_chars,
-                omitted_items=0,
-            ),
-            "preview": preview,
-        }
-        rendered = _finalize_envelope(envelope, max_bytes=max_bytes)
-        if len(rendered.encode("utf-8")) <= max_bytes:
-            return rendered
-        if preview_limit == 0:
-            return None
-        preview_limit //= 2
-
-
 _DELIVERY_BOUND_PRIORITY_KEYS: tuple[str, ...] = (
     "success",
     "receipt_id",
@@ -434,21 +189,6 @@ _DELIVERY_BOUND_DROPPABLE_KEYS: tuple[str, ...] = ("diagram",)
 # deprioritized keys remain *present* in the envelope — projected but
 # never dropped entirely.
 _DEPRIORITIZED_KEY_FLOOR_BYTES = 16
-
-
-def _serialized_string_prefix_length(value: str, max_bytes: int) -> int:
-    """Return the largest prefix whose serialized string body fits ``max_bytes``."""
-    low, high = 0, len(value)
-    best = 0
-    while low <= high:
-        midpoint = (low + high) // 2
-        serialized_bytes = len(_canonical_json(value[:midpoint]).encode("utf-8")) - 2
-        if serialized_bytes <= max_bytes:
-            best = midpoint
-            low = midpoint + 1
-        else:
-            high = midpoint - 1
-    return best
 
 
 def _tiered_projection(
@@ -935,367 +675,3 @@ def _spill_for_delivery_bound(
         return json.loads(rendered)
     except (ValueError, RecursionError):
         return {"success": False, "error": "response_budget_projection_invalid"}
-
-
-def enforce_response_budget(
-    result: Any,
-    *,
-    tool_name: str,
-    artifact_dir: Path | None,
-    config: OutputBudgetConfig,
-    force_spill: bool = False,
-    selected_result_token_limit: int | None = None,
-) -> Any:
-    """Return a bounded response of the same handler type.
-
-    Oversized content is atomically persisted before a projection is returned.
-    Artifact failure and missing-context cases fail closed without echoing the
-    original payload.
-
-    ``selected_result_token_limit`` is the authoritative bound selected for the
-    current downstream transport. For ordinary calls this is the backend's
-    unnegotiated limit; protected recipe calls may supply an attested limit.
-    Payloads whose estimated token count exceeds it are spilled even if they
-    pass the server-side exemption or response-byte ceilings.
-    """
-    segmented = _checkpoint_segmented_mapping(result)
-    if segmented is not None:
-        base, carrier = segmented
-        return _enforce_checkpoint_segmented_response(
-            result,
-            base=base,
-            carrier=carrier,
-            tool_name=tool_name,
-            artifact_dir=artifact_dir,
-            config=config,
-            force_spill=force_spill,
-            selected_result_token_limit=selected_result_token_limit,
-        )
-    try:
-        original = _serialized(result)
-    except (TypeError, ValueError, OverflowError, RecursionError):
-        return bounded_response_budget_failure(
-            result,
-            cause="serialization_failed",
-            tool_name=tool_name,
-            max_bytes=config.response_max_bytes,
-            original_utf8_bytes=0,
-        )
-    original_bytes = original.encode("utf-8")
-    original_size = len(original_bytes)
-    over_delivery_bound = (
-        selected_result_token_limit is not None
-        and selected_result_token_limit > 0
-        and _estimated_tokens(original_size) > selected_result_token_limit
-    )
-    exemption = RESPONSE_BACKSTOP_EXEMPTION_REGISTRY.get(tool_name)
-    if exemption is not None:
-        within_ceiling = (
-            len(original) <= exemption.max_chars and original_size <= exemption.max_utf8_bytes
-        )
-        if not within_ceiling:
-            return bounded_response_budget_failure(
-                result,
-                cause="exemption_ceiling_exceeded",
-                tool_name=tool_name,
-                max_bytes=config.response_max_bytes,
-                original_utf8_bytes=original_size,
-            )
-        if over_delivery_bound:
-            assert selected_result_token_limit is not None
-            return _spill_for_delivery_bound(
-                result,
-                tool_name=tool_name,
-                config=config,
-                artifact_dir=artifact_dir,
-                original=original,
-                original_size=original_size,
-                selected_result_token_limit=selected_result_token_limit,
-            )
-        _emit_response_budget_event(
-            "response_budget_exemption",
-            tool_name=_bounded_tool_name(tool_name),
-            measurement_id=exemption.measurement_id,
-            original_chars=len(original),
-            original_utf8_bytes=original_size,
-            max_chars=exemption.max_chars,
-            max_utf8_bytes=exemption.max_utf8_bytes,
-        )
-        return result
-    if (
-        not force_spill
-        and not over_delivery_bound
-        and len(original_bytes) <= config.response_max_bytes
-    ):
-        return result
-    if artifact_dir is None:
-        return bounded_response_budget_failure(
-            result,
-            cause="context_unavailable",
-            tool_name=tool_name,
-            max_bytes=config.response_max_bytes,
-            original_utf8_bytes=original_size,
-        )
-
-    path = _artifact_path(artifact_dir, tool_name)
-    try:
-        atomic_write(path, original)
-    except OSError:
-        return bounded_response_budget_failure(
-            result,
-            cause="artifact_publication_failed",
-            tool_name=tool_name,
-            max_bytes=config.response_max_bytes,
-            original_utf8_bytes=original_size,
-        )
-
-    published = str(path.resolve())
-    metadata = {
-        "artifact_path": published,
-        "sha256": hashlib.sha256(original_bytes).hexdigest(),
-        "original_utf8_bytes": len(original_bytes),
-    }
-
-    delivery_bound_bytes = (
-        CLIENT_CHARS_PER_TOKEN_POLICY.to_chars(TokenLimit(selected_result_token_limit)).value
-        if selected_result_token_limit is not None and selected_result_token_limit > 0
-        else None
-    )
-    projection_max_bytes = (
-        min(config.response_max_bytes, delivery_bound_bytes)
-        if delivery_bound_bytes is not None
-        else config.response_max_bytes
-    )
-    projection_inline_chars = min(config.inline_max_chars, projection_max_bytes)
-
-    parsed: Any = None
-    if isinstance(result, str):
-        try:
-            parsed = json.loads(result)
-        except (TypeError, ValueError, RecursionError):
-            parsed = None
-    else:
-        parsed = result
-
-    rendered: str | None = None
-    try:
-        if isinstance(parsed, dict):
-            rendered = _project_json_object(
-                parsed,
-                metadata=metadata,
-                max_bytes=projection_max_bytes,
-                inline_chars=projection_inline_chars,
-                delivery_bound_triggered=over_delivery_bound,
-            )
-        if rendered is None and not isinstance(parsed, dict):
-            rendered = _plain_spill_envelope(
-                original,
-                metadata=metadata,
-                max_bytes=projection_max_bytes,
-                inline_chars=projection_inline_chars,
-            )
-    except _ProjectionNonconvergentError as exc:
-        _emit_response_budget_event(
-            "response_budget_projection_nonconvergent",
-            tool_name=_bounded_tool_name(tool_name),
-            detail=str(exc),
-        )
-        rendered = bounded_response_budget_failure(
-            "",
-            cause="projection_nonconvergent",
-            tool_name=tool_name,
-            max_bytes=projection_max_bytes,
-            original_utf8_bytes=original_size,
-            artifact_path=published,
-        )
-    except RecursionError:
-        # Projection recurses per nesting level; the artifact is already
-        # persisted, so the recovery pointer must survive the stack failure.
-        rendered = bounded_response_budget_failure(
-            "",
-            cause="irreducible_shape",
-            tool_name=tool_name,
-            max_bytes=projection_max_bytes,
-            original_utf8_bytes=original_size,
-            artifact_path=published,
-        )
-    if rendered is None:
-        rendered = bounded_response_budget_failure(
-            "",
-            cause="irreducible_shape",
-            tool_name=tool_name,
-            max_bytes=projection_max_bytes,
-            original_utf8_bytes=original_size,
-            artifact_path=published,
-        )
-
-    assert isinstance(rendered, str)
-    _emit_response_budget_event(
-        "response_budget_spill",
-        tool_name=_bounded_tool_name(tool_name),
-        original_utf8_bytes=original_size,
-        projected_utf8_bytes=len(rendered.encode("utf-8")),
-    )
-    if isinstance(result, str):
-        return rendered
-    try:
-        return json.loads(rendered)
-    except (ValueError, RecursionError):
-        return {"success": False, "error": "response_budget_projection_invalid"}
-
-
-def _checkpoint_segmented_mapping(
-    result: Any,
-) -> tuple[dict[str, Any], dict[str, Any]] | None:
-    parsed: Any = result
-    if isinstance(result, str):
-        try:
-            parsed = json.loads(result)
-        except (TypeError, ValueError, RecursionError):
-            return None
-    if not isinstance(parsed, dict):
-        return None
-    carrier = parsed.get("recipe_segment")
-    if not isinstance(carrier, dict) or carrier.get("kind") not in {"success", "recovery"}:
-        return None
-    base = dict(parsed)
-    base.pop("recipe_segment")
-    return base, carrier
-
-
-def _segment_failure_like(
-    original: Any,
-    carrier: dict[str, Any],
-    *,
-    tool_name: str,
-) -> Any:
-    try:
-        failure = build_post_effect_segment_failure(carrier, tool_name=tool_name)
-    except RecipeSegmentDeliveryError as exc:
-        source_step = carrier.get("source_step", "")
-        logger.warning(
-            "recipe_segment_post_effect_carrier_dropped",
-            tool_name=tool_name,
-            source_step=source_step,
-            error=str(exc),
-        )
-        failure = {
-            "success": False,
-            "subtype": "recipe_segment_post_effect_delivery_failure",
-            "error": "Response shaping failed after the operation ran; do not repeat it.",
-            "tool_name": tool_name,
-            "step_name": source_step,
-            "operation_already_ran": True,
-            "do_not_repeat": True,
-        }
-    return _canonical_json(failure) if isinstance(original, str) else failure
-
-
-def post_effect_recipe_segment_failure(result: Any, *, tool_name: str) -> Any | None:
-    """Preserve the selected carrier when outer response enforcement itself fails."""
-    segmented = _checkpoint_segmented_mapping(result)
-    if segmented is None:
-        return None
-    _base, carrier = segmented
-    return _segment_failure_like(result, carrier, tool_name=tool_name)
-
-
-def _enforce_checkpoint_segmented_response(
-    original: Any,
-    *,
-    base: dict[str, Any],
-    carrier: dict[str, Any],
-    tool_name: str,
-    artifact_dir: Path | None,
-    config: OutputBudgetConfig,
-    force_spill: bool,
-    selected_result_token_limit: int | None,
-) -> Any:
-    """Artifactize only the base mapping and preserve the checkpoint carrier whole."""
-    combined = {**base, "recipe_segment": carrier}
-    try:
-        rendered = _canonical_json(combined)
-    except (TypeError, ValueError, OverflowError, RecursionError):
-        return _segment_failure_like(original, carrier, tool_name=tool_name)
-    rendered_size = len(rendered.encode("utf-8"))
-    final_limit = RECIPE_SEGMENT_MAX_BYTES - 1
-    over_delivery_bound = (
-        selected_result_token_limit is not None
-        and selected_result_token_limit > 0
-        and _estimated_tokens(rendered_size) > selected_result_token_limit
-    )
-    if not force_spill and not over_delivery_bound and rendered_size <= final_limit:
-        return rendered if isinstance(original, str) else combined
-    if artifact_dir is None:
-        return _segment_failure_like(original, carrier, tool_name=tool_name)
-
-    base_input: Any = _canonical_json(base) if isinstance(original, str) else base
-    shaped_base = enforce_response_budget(
-        base_input,
-        tool_name=tool_name,
-        artifact_dir=artifact_dir,
-        config=config,
-        force_spill=True,
-        selected_result_token_limit=selected_result_token_limit,
-    )
-    try:
-        shaped_mapping = json.loads(shaped_base) if isinstance(shaped_base, str) else shaped_base
-    except (TypeError, ValueError, RecursionError):
-        return _segment_failure_like(original, carrier, tool_name=tool_name)
-    if not isinstance(shaped_mapping, dict):
-        return _segment_failure_like(original, carrier, tool_name=tool_name)
-    shaped_error = shaped_mapping.get("error")
-    if isinstance(shaped_error, str) and shaped_error.startswith("response_budget_"):
-        return _segment_failure_like(original, carrier, tool_name=tool_name)
-
-    recombined = {**shaped_mapping, "recipe_segment": carrier}
-    try:
-        final_rendered = _canonical_json(recombined)
-    except (TypeError, ValueError, OverflowError, RecursionError):
-        return _segment_failure_like(original, carrier, tool_name=tool_name)
-    final_size = len(final_rendered.encode("utf-8"))
-    if final_size > final_limit or (
-        selected_result_token_limit is not None
-        and selected_result_token_limit > 0
-        and _estimated_tokens(final_size) > selected_result_token_limit
-    ):
-        return _segment_failure_like(original, carrier, tool_name=tool_name)
-    return final_rendered if isinstance(original, str) else recombined
-
-
-def shape_json_response(
-    payload: dict[str, Any],
-    *,
-    tool_name: str,
-    artifact_dir: Path,
-    config: OutputBudgetConfig,
-    selected_result_token_limit: int | None = None,
-) -> str:
-    """Serialize a tool dict, spilling once it crosses the source threshold."""
-    rendered = json.dumps(payload)
-    if len(rendered) <= config.inline_max_chars:
-        return rendered
-    shaped = enforce_response_budget(
-        rendered,
-        tool_name=tool_name,
-        artifact_dir=artifact_dir,
-        config=config,
-        force_spill=True,
-        selected_result_token_limit=selected_result_token_limit,
-    )
-    return shaped if isinstance(shaped, str) else json.dumps(shaped)
-
-
-__all__ = [
-    "RESPONSE_BUDGET_FAILURE_CAUSES",
-    "RESPONSE_SPILL_METADATA_KEY",
-    "RESPONSE_SPILL_METADATA_KEYS",
-    "RESPONSE_SPILL_REASONS",
-    "RESPONSE_SPILL_SCHEMA_DIGEST",
-    "RESPONSE_SPILL_SCHEMA_VERSION",
-    "bounded_response_budget_failure",
-    "emit_response_budget_failure",
-    "enforce_response_budget",
-    "post_effect_recipe_segment_failure",
-    "shape_json_response",
-]
